+++ src/motion/__tests__/delay.test.tsx
@@ -1,5 +1,5 @@
 import "../../../jest.setup"
-import { render } from "react-testing-library"
+import { render } from "@testing-library/react"
 import { motion } from ".."
 import * as React from "react"
 import { motionValue } from "../../value"
@@ -12,6 +12,44 @@ describe("delay attr", () => {
                 <motion.div
                     animate={{ x: 10 }}
                     transition={{ delay: 1, type: false }}
+                    style={{ x }}
+                />
+            )
+
+            const { rerender } = render(<Component />)
+            rerender(<Component />)
+
+            requestAnimationFrame(() => resolve(x.get()))
+        })
+
+        return expect(promise).resolves.toBe(0)
+    })
+    test("value-specific delay on instant transition", async () => {
+        const promise = new Promise(resolve => {
+            const x = motionValue(0)
+            const Component = () => (
+                <motion.div
+                    animate={{ x: 10 }}
+                    transition={{ x: { delay: 1, type: false } }}
+                    style={{ x }}
+                />
+            )
+
+            const { rerender } = render(<Component />)
+            rerender(<Component />)
+
+            requestAnimationFrame(() => resolve(x.get()))
+        })
+
+        return expect(promise).resolves.toBe(0)
+    })
+    test("value-specific delay on animation", async () => {
+        const promise = new Promise(resolve => {
+            const x = motionValue(0)
+            const Component = () => (
+                <motion.div
+                    animate={{ x: 10 }}
+                    transition={{ x: { delay: 1 } }}
                     style={{ x }}
                 />
             )